+++ example/ios/EventideTests/ReminderTests.swift
@@ -24,7 +24,7 @@ final class ReminderTests: XCTestCase {
                     title: "title",
                     color: UIColor.red,
                     isWritable: true,
-                    account: Account(name: "local", type: "local"),
+                    account: Account(id: "local", name: "local", type: "local"),
                     events: [
                         MockEvent(
                             id: "1",
@@ -73,7 +73,7 @@ final class ReminderTests: XCTestCase {
                     title: "title",
                     color: UIColor.red,
                     isWritable: true,
-                    account: Account(name: "local", type: "local"),
+                    account: Account(id: "local", name: "local", type: "local"),
                     events: [
                         MockEvent(
                             id: "1",
@@ -123,7 +123,7 @@ final class ReminderTests: XCTestCase {
                     title: "title",
                     color: UIColor.red,
                     isWritable: true,
-                    account: Account(name: "local", type: "local"),
+                    account: Account(id: "local", name: "local", type: "local"),
                     events: [
                         MockEvent(
                             id: "2",
@@ -173,7 +173,7 @@ final class ReminderTests: XCTestCase {
                     title: "title",
                     color: UIColor.red,
                     isWritable: true,
-                    account: Account(name: "local", type: "local"),
+                    account: Account(id: "local", name: "local", type: "local"),
                     events: [
                         MockEvent(
                             id: "1",
@@ -220,7 +220,7 @@ final class ReminderTests: XCTestCase {
                     title: "title",
                     color: UIColor.red,
                     isWritable: true,
-                    account: Account(name: "local", type: "local"),
+                    account: Account(id: "local", name: "local", type: "local"),
                     events: [
                         MockEvent(
                             id: "2",
@@ -271,7 +271,7 @@ final class ReminderTests: XCTestCase {
                     title: "title",
                     color: UIColor.red,
                     isWritable: true,
-                    account: Account(name: "local", type: "local"),
+                    account: Account(id: "local", name: "local", type: "local"),
                     events: [
                         MockEvent(
                             id: "1",
@@ -324,7 +324,7 @@ final class ReminderTests: XCTestCase {
                     title: "title",
                     color: UIColor.red,
                     isWritable: true,
-                    account: Account(name: "local", type: "local"),
+                    account: Account(id: "local", name: "local", type: "local"),
                     events: [
                         MockEvent(
                             id: "1",
@@ -374,7 +374,7 @@ final class ReminderTests: XCTestCase {
                     title: "title",
                     color: UIColor.red,
                     isWritable: true,
-                    account: Account(name: "local", type: "local"),
+                    account: Account(id: "local", name: "local", type: "local"),
                     events: [
                         MockEvent(
                             id: "1",
@@ -427,7 +427,7 @@ final class ReminderTests: XCTestCase {
                     title: "title",
                     color: UIColor.red,
                     isWritable: true,
-                    account: Account(name: "local", type: "local"),
+                    account: Account(id: "local", name: "local", type: "local"),
                     events: [
                         MockEvent(
                             id: "1",
@@ -476,7 +476,7 @@ final class ReminderTests: XCTestCase {
                     title: "title",
                     color: UIColor.red,
                     isWritable: true,
-                    account: Account(name: "local", type: "local"),
+                    account: Account(id: "local", name: "local", type: "local"),
                     events: [
                         MockEvent(
                             id: "1",